Silenced some more Compiler warnings.
git-svn-id: svn://scribus.net/trunk/Scribus@17810 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/gettext/sxwim/contentreader.cpp
@@ -127,7 +127,7 @@ bool ContentReader::startElement(const QString&, const QString&, const QString &
 	{
 		bool isOrdered = false;
 		int levelIndex = listLevel - 1;
-		if (levelIndex >= 0 && levelIndex < isOrdered2.size())
+		if (levelIndex >= 0 && levelIndex < static_cast<int>(isOrdered2.size()))
 			isOrdered = isOrdered2[listLevel - 1];
 		if (isOrdered)
 		{
